requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
google-cloud-vision==3.4.5
opencv-python-headless==4.8.1.78
numpy==1.23.3
Pillow==9.2.0
opencv-python==4.5.5.64
